Adding Torch Prefab, and SoundManager game object
Adding Torch for the player which has a battery that decreases automatically as long as the torch is switched on, however, battery screen design is not made yet. Moreover, SoundManager gameobject is used to map all of our game sounds to it, it has an attached SoundManagerScript which we will use in our scripts to play sounds, rather than creating an audio source for each object we use.